port to fabric 1.20.x

+++ src/main/java/bottomtextdanny/particular/config/SquidInkSplashConfig.java
@@ -17,7 +17,7 @@ public class SquidInkSplashConfig {
 	}
 
 	public double ejectionStrength() {
-		return ejectionStrength.get().doubleValue();
+		return ejectionStrength.get();
 	}
 
 	public int min() {
